31 | Comment 1:N relationship

+++ articles/views.py
@@ -1,11 +1,13 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from IPython import embed
-from .models import Article, Comment
-from .forms import ArticleForm, CommentForm
 from django.views.decorators.http import require_POST # POST여야만 동작할 수 있다
 from django.contrib.auth.decorators import login_required # login해야만 동작할 수 있게 하는 데코레이터
 from django.contrib import messages
+from django.core.exceptions import PermissionDenied
+from django.http import HttpResponseForbidden
+from .models import Article, Comment
+from .forms import ArticleForm, CommentForm
 
+from IPython import embed
 
 # Create your views here.
 def index(request):
@@ -56,10 +58,13 @@ def create(request):
 def detail(request, article_pk):
     article = Article.objects.get(pk=article_pk)
     comments = article.comment_set.all()
+    comment_form = CommentForm()
+    re = request.user
+    print(re)
     context = {
         'article': article,
         'comments': comments,
-        'request_user': request.user
+        'comment_form': comment_form
     }
     return render(request, 'articles/detail.html', context)
 
@@ -78,8 +83,7 @@ def delete(request, article_pk):
         article.delete()
         return redirect('articles:index')
     else:
-        messages.add_message(request, messages.INFO, '삭제 권한이 없습니다.')
-        return render('articles/index.html')
+        return HttpResponseForbidden
 
 def edit(request, article_pk):
     article = Article.objects.get(pk=article_pk)
@@ -88,25 +92,30 @@ def edit(request, article_pk):
     }
     return render(request, 'articles/edit.html', context)
 
+
 def update(request, article_pk):
     article = Article.objects.get(pk=article_pk)
-    if request.method == 'POST':
-        article_form = ArticleForm(request.POST, instance=article)
-        if article_form.is_valid():
-            article = article_form.save()
-            # article.title = article_form.cleaned_data.get('title')
-            # article.content = article_form.cleaned_data.get('content')
-            # article.save()
-            return redirect('articles:detail', article.pk)
-    else:
-        article_form = ArticleForm(instance=article)
+    if article.user == request.user: 
+        if request.method == 'POST':
+            article_form = ArticleForm(request.POST, instance=article)
+            if article_form.is_valid():
+                article = article_form.save()
+                # article.title = article_form.cleaned_data.get('title')
+                # article.content = article_form.cleaned_data.get('content')
+                # article.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            article_form = ArticleForm(instance=article)
 
-    context = {
-        'article_form': article_form
-    }
-    return render(request, 'articles/form.html', context)
+        context = {
+            'article_form': article_form
+        }
+        return render(request, 'articles/form.html', context)
+    else:
+        return HttpResponseForbidden
 
 @require_POST
+@login_required
 def comment_create(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     # 1. modelform에 사용자 입력값 넣고
@@ -117,7 +126,6 @@ def comment_create(request, article_pk):
         comment.article = article # 내가 직접 조작한후
         comment.user = request.user
         comment.save() # DB에 쿼리날린다!
-
     else:
         messages.success(request, '댓글이 형식에 맞지 않습니다.')
         
@@ -125,8 +133,12 @@ def comment_create(request, article_pk):
 
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
-    if request.method == 'POST':
-        comment = Comment.objects.get(pk=comment_pk)
-        comment.delete()
-        messages.add_message(request, messages.INFO, '댓글이 삭제되었습니다.')
-        return redirect('articles:detail', article_pk)
+    comment = Comment.objects.get(pk=comment_pk)
+    if comment.user == request.user:
+        if request.method == 'POST':
+            comment.delete()
+            messages.add_message(request, messages.INFO, '댓글이 삭제되었습니다.')
+            return redirect('articles:detail', article_pk)
+    else:
+        # raise PermissionDenied
+        return HttpResponseForbidden # 위에거랑 이러랑 둘다 가능!
